requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
fastapi==0.118.2
Flask==3.1.2
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.3
numpy==2.3.3
openai==2.2.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
srt==3.5.3
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
vosk==0.3.45
websockets==15.0.1
Werkzeug==3.1.3
